Add seperate class for AddCommand, GoCommand and BackCommand
