+++ src/main/java/commands/GoChapterCommand.java
@@ -1,0 +1,59 @@
+package commands;
+
+import access.Access;
+import exception.IncorrectAccessLevelException;
+import manager.card.Card;
+import manager.chapter.Chapter;
+import manager.module.ChapterList;
+import storage.Storage;
+import ui.Ui;
+
+import java.io.FileNotFoundException;
+import java.util.ArrayList;
+
+public class GoChapterCommand extends GoCommand {
+
+    private String chapter;
+
+    public GoChapterCommand(String chapter) {
+        this.chapter = chapter;
+    }
+
+    @Override
+    public void execute(Ui ui, Access access, Storage storage) throws IncorrectAccessLevelException {
+        String result = goChapter(access, storage);
+        if (result.equals("")) {
+            return;
+        }
+        ui.showToUser(result);
+    }
+
+    private String goChapter(Access access, Storage storage) throws IncorrectAccessLevelException {
+        boolean isLevelExist = false;
+        String result = "";
+        String currentChapter = this.chapter;
+        ChapterList chapters = access.getModule().getChapters();
+        ArrayList<Chapter> allChapters = chapters.getAllChapters();
+        for (Chapter chapter : allChapters) {
+            if (currentChapter.equalsIgnoreCase(chapter.getChapterName())) {
+                access.setChapterLevel(currentChapter);
+                isLevelExist = true;
+                try {
+                    ArrayList<Card> allCards = storage.loadCard(access.getModuleLevel(), chapter.getChapterName());
+                    if (allCards.size() == 0) {
+                        result = "This is a new chapter, you can try to add flashcards inside!";
+                    }
+                    chapter.setCards(allCards);
+                    access.setChapter(chapter);
+                } catch (FileNotFoundException e) {
+                    result = "The chapter file cannot be found.";
+                }
+                break;
+            }
+        }
+        if (!isLevelExist) {
+            result = "Sorry, I cannot find this chapter, please add this chapter first";
+        }
+        return result;
+    }
+}
