+++ src/main/java/parser/Parser.java
@@ -2,28 +2,7 @@ package parser;
 
 import access.Access;
 
-import commands.AddCommand;
-import commands.BackCommand;
-import commands.Command;
-import commands.EditCommand;
-import commands.ExcludeCommand;
-import commands.ExitCommand;
-import commands.GoCommand;
-import commands.HelpCommand;
-import commands.HistoryCommand;
-import commands.ListCommand;
-import commands.ListDueCommand;
-import commands.PreviewCommand;
-import commands.RateCommand;
-import commands.RemoveCardCommand;
-import commands.RemoveChapterCommand;
-import commands.RemoveCommand;
-import commands.RemoveModuleCommand;
-import commands.RescheduleCommand;
-import commands.ReviseCommand;
-import commands.ExcludeCommand;
-import commands.ShowRateCommand;
-
+import commands.*;
 import exception.IncorrectAccessLevelException;
 import exception.InvalidFileFormatException;
 import exception.InvalidInputException;
@@ -31,6 +10,7 @@ import storage.Storage;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
+import static common.Messages.ADMIN;
 import static common.Messages.CARD;
 import static common.Messages.CHAPTER;
 import static common.Messages.MESSAGE_EXTRA_ARGS;
@@ -45,10 +25,6 @@ public class Parser {
     private static final String QUESTION_ANSWER_PREFIX = " \\| ";
     private static final String QUESTION_PREFIX = "q:";
     private static final String ANSWER_PREFIX = "a:";
-
-    private static final String ADMIN_LEVEL = "admin";
-    private static final String MODULE_LEVEL = "module";
-    private static final String CHAPTER_LEVEL = "chapter";
 
     public static Command parse(String fullCommand, Access access)
             throws InvalidInputException, IncorrectAccessLevelException {
@@ -74,9 +50,9 @@ public class Parser {
         case EditCommand.COMMAND_WORD:
             return prepareEdit(commandArgs, access);
         case BackCommand.COMMAND_WORD:
-            return prepareBack(commandArgs);
+            return prepareBack(commandArgs, access);
         case GoCommand.COMMAND_WORD:
-            return prepareGo(commandArgs);
+            return prepareGo(commandArgs, access);
         case ListDueCommand.COMMAND_WORD:
             return prepareListDue(commandArgs);
         case HistoryCommand.COMMAND_WORD:
@@ -136,20 +112,42 @@ public class Parser {
         return new HistoryCommand(commandArgs);
     }
 
-    private static Command prepareGo(String commandArgs) throws InvalidInputException {
+    private static Command prepareGo(String commandArgs, Access access)
+            throws InvalidInputException, IncorrectAccessLevelException {
+        String level = getLevel(access);
         if (commandArgs.isEmpty()) {
             throw new InvalidInputException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
                     GoCommand.COMMAND_WORD) + GoCommand.MESSAGE_USAGE);
         }
-        return new GoCommand(commandArgs);
+
+        switch (level) {
+        case ADMIN:
+            return new GoModuleCommand(commandArgs);
+        case MODULE:
+            return new GoChapterCommand(commandArgs);
+        default:
+            throw new IncorrectAccessLevelException("Go command can only be called "
+                    + "at admin and module level.");
+        }
     }
 
-    private static Command prepareBack(String commandArgs) throws InvalidInputException {
+    private static Command prepareBack(String commandArgs, Access access)
+            throws InvalidInputException, IncorrectAccessLevelException {
+        String level = getLevel(access);
         if (!commandArgs.isEmpty()) {
             throw new InvalidInputException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
                     BackCommand.COMMAND_WORD) + BackCommand.MESSAGE_USAGE);
         }
-        return new BackCommand();
+
+        switch (level) {
+        case CHAPTER:
+            return new BackModuleCommand();
+        case MODULE:
+            return new BackAdminCommand();
+        default:
+            throw new IncorrectAccessLevelException("Back command can only be called "
+                    + "at module and chapter level.");
+        }
     }
 
     private static String[] splitCommandTypeAndArgs(String userCommand) {
@@ -170,25 +168,23 @@ public class Parser {
 
     private static Command prepareAdd(String commandArgs, Access access)
             throws InvalidInputException, IncorrectAccessLevelException {
-        if (access.isChapterLevel()) {
+        String level = getLevel(access);
+        switch (level) {
+        case ADMIN:
             if (commandArgs.isEmpty()) {
-                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddCommand.CARD_MESSAGE_USAGE);
-            }
-            return prepareAddCard(commandArgs);
-        } else if (access.isModuleLevel()) {
-            if (commandArgs.isEmpty()) {
-                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddCommand.CHAPTER_MESSAGE_USAGE);
-            }
-            return prepareAddChapter(commandArgs);
-        } else if (access.isAdminLevel()) {
-            if (commandArgs.isEmpty()) {
-                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddCommand.MODULE_MESSAGE_USAGE);
+                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddModuleCommand.MODULE_MESSAGE_USAGE);
             }
             return prepareAddModule(commandArgs);
-        } else {
-            assert !access.isChapterLevel() && !access.isAdminLevel() && !access.isModuleLevel() : access.getLevel();
-            throw new IncorrectAccessLevelException(String.format(MESSAGE_INCORRECT_ACCESS,
-                    AddCommand.COMMAND_WORD));
+        case MODULE:
+            if (commandArgs.isEmpty()) {
+                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddChapterCommand.CHAPTER_MESSAGE_USAGE);
+            }
+            return prepareAddChapter(commandArgs);
+        default:
+            if (commandArgs.isEmpty()) {
+                throw new InvalidInputException(MESSAGE_MISSING_ARGS + AddCardCommand.CARD_MESSAGE_USAGE);
+            }
+            return prepareAddCard(commandArgs);
         }
     }
 
@@ -199,21 +195,37 @@ public class Parser {
             String answer = parseAnswer(args[1]);
             if (question.isEmpty() || answer.isEmpty()) {
                 throw new InvalidInputException("The content for question / answer is empty.\n"
-                        + AddCommand.CARD_MESSAGE_USAGE);
+                        + AddCardCommand.CARD_MESSAGE_USAGE);
             }
-            return new AddCommand(question, answer, CHAPTER_LEVEL);
+            return new AddCardCommand(question, answer);
         } catch (IndexOutOfBoundsException e) {
             throw new InvalidInputException("The format for the add command is incorrect.\n"
-                    + AddCommand.CARD_MESSAGE_USAGE);
+                    + AddCardCommand.CARD_MESSAGE_USAGE);
         }
     }
 
     private static Command prepareAddChapter(String commandArgs) {
-        return new AddCommand(commandArgs, MODULE_LEVEL);
+        return new AddChapterCommand(commandArgs);
     }
 
     private static Command prepareAddModule(String commandArgs) {
-        return new AddCommand(commandArgs, ADMIN_LEVEL);
+        return new AddModuleCommand(commandArgs);
+    }
+
+    private static String getLevel(Access access) throws IncorrectAccessLevelException {
+        String level;
+        if (access.isAdminLevel()) {
+            level = ADMIN;
+        } else if (access.isModuleLevel()) {
+            level = MODULE;
+        } else if (access.isChapterLevel()) {
+            level = CHAPTER;
+        } else {
+            assert !access.isChapterLevel() && !access.isAdminLevel() && !access.isModuleLevel() : access.getLevel();
+            throw new IncorrectAccessLevelException(String.format(MESSAGE_INCORRECT_ACCESS,
+                    AddCommand.COMMAND_WORD));
+        }
+        return level;
     }
 
     private static Command prepareRemove(String commandArgs, Access access) throws InvalidInputException,
@@ -300,7 +312,7 @@ public class Parser {
                         + EditCommand.CARD_MESSAGE_USAGE);
             }
 
-            return new EditCommand(editIndex, question, answer, CHAPTER_LEVEL);
+            return new EditCommand(editIndex, question, answer, CHAPTER);
         } catch (NumberFormatException e) {
             throw new InvalidInputException("The flashcard number needs to be an integer.\n"
                     + EditCommand.CARD_MESSAGE_USAGE);
@@ -329,7 +341,7 @@ public class Parser {
             }
 
             int editIndex = Integer.parseInt(args[0].trim()) - 1;
-            return new EditCommand(editIndex, args[1].trim().toLowerCase(), ADMIN_LEVEL);
+            return new EditCommand(editIndex, args[1].trim().toLowerCase(), ADMIN);
         } catch (NumberFormatException e) {
             throw new InvalidInputException("The module number needs to be an integer.\n"
                     + EditCommand.MODULE_MESSAGE_USAGE);
@@ -358,7 +370,7 @@ public class Parser {
             }
 
             int editIndex = Integer.parseInt(args[0].trim()) - 1;
-            return new EditCommand(editIndex, args[1].trim().toLowerCase(), MODULE_LEVEL);
+            return new EditCommand(editIndex, args[1].trim().toLowerCase(), MODULE);
         } catch (NumberFormatException e) {
             throw new InvalidInputException("The chapter number needs to be an integer.\n"
                     + EditCommand.CHAPTER_MESSAGE_USAGE);
